requirements.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-queryfields==1.0.0
filelock==3.7.1
h11==0.13.0
idna==3.3
linkedin-api==1.1.0
lxml==4.9.0
mysqlclient==2.1.0
numpy==1.22.4
outcome==1.2.0
pandas==1.4.2
Pillow==9.1.1
platformdirs==2.5.2
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
wsproto==1.1.0
